+++ API/src/routes/projects/projects.ts
@@ -186,7 +186,7 @@ router.get("/:id", authenticate, (req: Request, res: Response): void => {
 router.post(
   "/",
   authenticate,
-  // authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR, ROLES.PROGRAM_MANAGER]),
+  authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR, ROLES.PROGRAM_MANAGER]),
   (req: Request, res: Response): void => {
     projectsController.createProject(req, res);
   }
@@ -247,7 +247,7 @@ router.post(
 router.put(
   "/:id",
   authenticate,
-  // authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR, ROLES.PROGRAM_MANAGER]),
+  authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR, ROLES.PROGRAM_MANAGER]),
   (req: Request, res: Response): void => {
     projectsController.updateProject(req, res);
   }
@@ -290,7 +290,7 @@ router.put(
 router.delete(
   "/:id",
   authenticate,
-  // authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
+  authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
   (req: Request, res: Response): void => {
     projectsController.deleteProject(req, res);
   }
@@ -359,7 +359,7 @@ router.get(
 router.post(
   "/:projectId/users",
   authenticate,
-  // authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
+  authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
   (req: Request, res: Response): void => {
     projectsController.assignments.assignUserToProject(req, res);
   }
@@ -393,7 +393,7 @@ router.post(
 router.delete(
   "/:projectId/users/:userId",
   authenticate,
-  // authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
+  authorize([ROLES.SUPER_ADMIN, ROLES.SYSTEM_ADMINISTRATOR]),
   (req: Request, res: Response): void => {
     projectsController.assignments.removeUserFromProject(req, res);
   }
